feat: improve Dataset Actions detail, edit and list views
Adds missing widgets to edit and detail views showing which Organisation
was affected by an action and which person in the CRM made the change.  This
fixes https://github.com/IATI/iati-registry-suitecrm-extension/issues/3.
As part of this a fix was applied to the underlying database structure.  The
list view was also augmented with additional fields.