feat: buttons to open Validator and Dashboard from dataset views
Adds buttons in the dataset list view, and action menu links in the
dataset detail view, to open the relevant Validator and Dashboard pages.
These are correct for the validator but since the Dashboard pages are not
yet created these are placeholders that take the user to the Dashboard
home page.